+++ apps/native/components/ui/BaseBottomSheet.tsx
@@ -1,0 +1,216 @@
+import React, {
+  forwardRef,
+  useCallback,
+  useEffect,
+  useImperativeHandle,
+  useRef,
+  type ReactNode,
+} from 'react';
+import {
+  Modal,
+  View,
+  Text,
+  TouchableWithoutFeedback,
+  TouchableOpacity,
+  StyleSheet,
+  Animated,
+  PanResponder,
+  useWindowDimensions,
+} from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import { Colors } from '@/constants/Colors';
+import { useColorScheme } from '@/hooks/useColorScheme';
+
+export interface BaseBottomSheetRef {
+  close: () => void;
+}
+
+interface BaseBottomSheetProps {
+  visible: boolean;
+  title: string;
+  onClose: () => void;
+  heightRatio?: number;
+  children: ReactNode;
+}
+
+const DRAG_THRESHOLD = 80;
+
+export const BaseBottomSheet = forwardRef<BaseBottomSheetRef, BaseBottomSheetProps>(
+  function BaseBottomSheet({ visible, title, onClose, heightRatio = 0.4, children }, ref) {
+  const colorScheme = useColorScheme();
+  const colors = Colors[colorScheme ?? 'light'];
+  const insets = useSafeAreaInsets();
+  const { height: screenHeight } = useWindowDimensions();
+  const sheetHeight = screenHeight * heightRatio;
+
+  const translateY = useRef(new Animated.Value(sheetHeight)).current;
+  const opacity = useRef(new Animated.Value(0)).current;
+  const onCloseRef = useRef(onClose);
+  onCloseRef.current = onClose;
+
+  const handleClose = useCallback(() => {
+    Animated.parallel([
+      Animated.timing(translateY, {
+        toValue: sheetHeight,
+        duration: 230,
+        useNativeDriver: true,
+      }),
+      Animated.timing(opacity, {
+        toValue: 0,
+        duration: 230,
+        useNativeDriver: true,
+      }),
+    ]).start(() => onCloseRef.current());
+  }, [translateY, opacity, sheetHeight]);
+
+  useImperativeHandle(ref, () => ({ close: handleClose }), [handleClose]);
+
+  const panResponder = useRef(
+    PanResponder.create({
+      onStartShouldSetPanResponder: () => true,
+      onMoveShouldSetPanResponder: (_, gestureState) =>
+        Math.abs(gestureState.dy) > 5,
+      onPanResponderMove: (_, gestureState) => {
+        if (gestureState.dy > 0) {
+          translateY.setValue(gestureState.dy);
+        }
+      },
+      onPanResponderRelease: (_, gestureState) => {
+        if (gestureState.dy > DRAG_THRESHOLD) {
+          handleClose();
+        } else {
+          Animated.spring(translateY, {
+            toValue: 0,
+            useNativeDriver: true,
+            tension: 65,
+            friction: 11,
+          }).start();
+        }
+      },
+    })
+  ).current;
+
+  useEffect(() => {
+    if (visible) {
+      Animated.parallel([
+        Animated.spring(translateY, {
+          toValue: 0,
+          useNativeDriver: true,
+          tension: 65,
+          friction: 11,
+        }),
+        Animated.timing(opacity, {
+          toValue: 1,
+          duration: 300,
+          useNativeDriver: true,
+        }),
+      ]).start();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [visible]);
+
+  return (
+    <Modal
+      visible={visible}
+      transparent
+      animationType="none"
+      statusBarTranslucent
+      onRequestClose={handleClose}
+    >
+      <View style={styles.container}>
+        <TouchableWithoutFeedback onPress={handleClose}>
+          <Animated.View
+            style={[
+              styles.backdrop,
+              {
+                opacity: opacity.interpolate({
+                  inputRange: [0, 1],
+                  outputRange: [0, 0.5],
+                }),
+              },
+            ]}
+          />
+        </TouchableWithoutFeedback>
+
+        <Animated.View
+          style={[
+            styles.sheet,
+            {
+              maxHeight: sheetHeight,
+              backgroundColor: colors.background,
+              paddingBottom: insets.bottom + 24,
+              transform: [{ translateY }],
+            },
+          ]}
+        >
+          <View {...panResponder.panHandlers} style={styles.handle}>
+            <View
+              style={[
+                styles.handleBar,
+                { backgroundColor: colors.textDisabled },
+              ]}
+            />
+          </View>
+
+          <View style={styles.header}>
+            <Text style={[styles.title, { color: colors.text }]}>{title}</Text>
+            <TouchableOpacity onPress={handleClose} style={styles.closeButton}>
+              <Ionicons name="close" size={24} color={colors.text} />
+            </TouchableOpacity>
+          </View>
+
+          {children}
+        </Animated.View>
+      </View>
+    </Modal>
+  );
+  },
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  backdrop: {
+    ...StyleSheet.absoluteFillObject,
+    backgroundColor: 'black',
+  },
+  sheet: {
+    position: 'absolute',
+    left: 0,
+    right: 0,
+    bottom: 0,
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: -2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 10,
+    elevation: 5,
+  },
+  handle: {
+    alignItems: 'center',
+    paddingVertical: 12,
+  },
+  handleBar: {
+    width: 40,
+    height: 4,
+    borderRadius: 2,
+  },
+  header: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    paddingHorizontal: 24,
+    paddingBottom: 12,
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: '600',
+    letterSpacing: -0.5,
+  },
+  closeButton: {
+    padding: 4,
+  },
+});
